TЗ 6, первое ревью. Реализация класса MyLinkedList. Добавление новой функциональности в InMemoryHistoryManager. Покрытие новыми тестами.
