+++ src/ru/yandex/Practicum/manager/InMemoryHistoryManager.java
@@ -1,12 +1,9 @@
 package ru.yandex.Practicum.manager;
 
 import ru.yandex.Practicum.model.Task;
-import ru.yandex.Practicum.util.Node;
+import ru.yandex.Practicum.util.MyLinkedList;
 
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 public class InMemoryHistoryManager implements HistoryManager {
 
@@ -25,71 +22,5 @@ public class InMemoryHistoryManager implements HistoryManager {
     @Override
     public void remove(int id) {
         myLinkedList.removeNode(id);
-    }
-}
-
-class MyLinkedList<E extends Task> {
-    Node<E> first;
-    Node<E> last;
-    private final Map<Integer, Node<E>> history = new HashMap<>();
-
-    void linkLast(E e) {
-        if (history.containsKey(e.getId())) {
-            removeNode(e.getId());
-        }
-
-        final Node<E> l = last;
-        final Node<E> newNode = new Node<>(l, e, null);
-        last = newNode;
-        if (l == null)
-            first = newNode;
-        else
-            l.next = newNode;
-
-        history.put(e.getId(), newNode);
-    }
-
-    public void unlink(int id) {
-        Node<E> x = history.get(id);
-        final Node<E> next = x.next;
-        final Node<E> prev = x.prev;
-
-        if (prev == null) {
-            first = next;
-        } else {
-            prev.next = next;
-            x.prev = null;
-        }
-
-        if (next == null) {
-            last = prev;
-        } else {
-            next.prev = prev;
-            x.next = null;
-        }
-        x.e = null;
-        history.remove(id);
-    }
-
-    public void removeNode(int id) {
-        unlink(id);
-    }
-
-    public List<E> getTasks() {
-        List<E> historyList = new ArrayList<>();
-        Node<E> item = first;
-        while (item != null) {
-            historyList.add(item.e);
-            item = item.next;
-        }
-        return historyList;
-    }
-
-    public Map<Integer, Node<E>> getHashMap() {
-        return history;
-    }
-
-    public Node<E> getNode(int id) {
-        return history.get(id);
     }
 }
